fix clean up - use separate file for customer endpoints
